requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
distro==1.7.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
tabula-py==2.5.1
urllib3==1.26.12
